+++ pkg/app/run_test.go
@@ -80,7 +80,6 @@ profile
 secret-store
 secret-store-entry
 service
-service-auth
 stats
 tls-config
 tls-custom