+++ pkg/commands/service/auth/list.go
@@ -1,4 +1,4 @@
-package serviceauth
+package auth
 
 import (
 	"context"